#126 for Adam Voysey - Update Thread Utils (to match UM) Merged into /shumlib/trunk: /shumlib/branches/dev/adamvoysey/r1976_thread_utils_update@2110 cf. /shumlib/trunk@1976
git-svn-id: https://metomi/svn/utils.xm/shumlib/trunk@2136 9980b737-bc2f-46b8-919e-b198a3dff10a

+++ shum_data_conv/src/c_shum_data_conv.c
@@ -1158,7 +1158,7 @@ if(format == IBM32FLT_FORMAT || format == IBM64FLT_FORMAT ){
 
   expo = in_expo/4 + (in_expo >= 0);
 
-  expo_diff = (4 - ((uint64_t)in_expo & 3));
+  expo_diff = (size_t)(4 - ((uint64_t)in_expo & 3));
   // equivalent to expo_diff = (in_expo>=0)*4-(in_expo%4) on
   // a two's compliment system.
   // C99 standard requires two's compliment representation for fixed width
